+++ cmd/tfgateway/redis.go
@@ -1,0 +1,52 @@
+package tfgateway
+
+import (
+	"fmt"
+	"net/url"
+	"time"
+
+	"github.com/gomodule/redigo/redis"
+)
+
+func newRedisPool(address string) (*redis.Pool, error) {
+	u, err := url.Parse(address)
+	if err != nil {
+		return nil, err
+	}
+	var host string
+	switch u.Scheme {
+	case "tcp":
+		host = u.Host
+	case "unix":
+		host = u.Path
+	default:
+		return nil, fmt.Errorf("unknown scheme '%s' expecting tcp or unix", u.Scheme)
+	}
+	var opts []redis.DialOption
+
+	if u.User != nil {
+		opts = append(
+			opts,
+			redis.DialPassword(u.User.Username()),
+		)
+	}
+
+	return &redis.Pool{
+		Dial: func() (redis.Conn, error) {
+			return redis.Dial(u.Scheme, host, opts...)
+		},
+		TestOnBorrow: func(c redis.Conn, t time.Time) error {
+			if time.Since(t) > 10*time.Second {
+				//only check connection if more than 10 second of inactivity
+				_, err := c.Do("PING")
+				return err
+			}
+
+			return nil
+		},
+		MaxActive:   3,
+		MaxIdle:     3,
+		IdleTimeout: 1 * time.Minute,
+		Wait:        true,
+	}, nil
+}
