+++ delegate.go
@@ -20,7 +20,7 @@ func (p *Provisioner) domainDeleateProvision(ctx context.Context, r *provision.R
 	}
 	log.Info().Str("id", r.ID).Msgf("provision Delegate %+v", data)
 
-	return nil, p.dns.AddDomainDelagate(r.User, data)
+	return nil, p.dns.AddDomainDelagate(r.User, data.Domain)
 }
 
 func (p *Provisioner) domainDeleateDecomission(ctx context.Context, r *provision.Reservation) error {
@@ -30,5 +30,5 @@ func (p *Provisioner) domainDeleateDecomission(ctx context.Context, r *provision
 	}
 	log.Info().Str("id", r.ID).Msgf("decomission Delegate %+v", data)
 
-	return p.dns.RemoveDomainDelagate(r.User, data)
+	return p.dns.RemoveDomainDelagate(r.User, data.Domain)
 }
